requirements.txt
langchain
faiss-cpu
google-generativeai
flask
